+++ loqusdb/utils/variant.py
@@ -3,12 +3,13 @@ import logging
 from pprint import pprint as pp
 LOG = logging.getLogger(__name__)
 
-def format_info(variant, variant_type='snv'):
+def format_info(variant, variant_type='snv', nr_cases=None, add_freq=False):
     """Format the info field for SNV variants
     
     Args:
         variant(dict)
         variant_type(str): snv or sv
+        nr_cases(int)
     
     Returns:
         vcf_info(str): A VCF formated info field
@@ -19,12 +20,15 @@ def format_info(variant, variant_type='snv'):
     homozygotes = variant.get('homozygote')
     hemizygotes = variant.get('hemizygote')
 
-    
     vcf_info = f"Obs={observations}"
     if homozygotes:
         vcf_info += f";Hom={homozygotes}"
     if hemizygotes:
         vcf_info += f";Hem={hemizygotes}"
+
+    if add_freq and nr_cases and nr_cases > 0:
+        frequency = observations / nr_cases
+        vcf_info += f";Frq={frequency:.5f}"
 
     # This is SV specific
     if variant_type == 'sv':
@@ -34,12 +38,13 @@ def format_info(variant, variant_type='snv'):
 
     return vcf_info
 
-def format_variant(variant, variant_type='snv'):
+def format_variant(variant, variant_type='snv', nr_cases=None, add_freq=False):
     """Convert variant information to a VCF formated string
     
     Args:
         variant(dict)
         variant_type(str)
+        nr_cases(int)
     
     Returns:
         vcf_variant(str)
@@ -57,7 +62,7 @@ def format_variant(variant, variant_type='snv'):
 
     info = None
     
-    info = format_info(variant, variant_type=variant_type)
+    info = format_info(variant, variant_type=variant_type, nr_cases=nr_cases, add_freq=add_freq)
 
     variant_line = f"{chrom}\t{pos}\t.\t{ref}\t{alt}\t.\t.\t{info}"